issue #153 - unify getCbaState helper
