+++ src/js/ui/projects.js
@@ -25,7 +25,8 @@ const {NO_PROJ_SELECTED, NO_PROJ_GROUP_SELECTED, NO_ACTION_SELECTED,
 const projectsDb = require("../db/projects");
 const prefs = require("../db/prefs");
 const ActionInputs = require("./ActionInputs");
-const {sendRpcMessage, addRpcListener, removeRpcListener} = require("../rpc/client");
+const {sendRpcMessage} = require("../rpc/client");
+const getCbaState = require("../commons/getCbaState");
 
 /**
  * @typedef {import("../rpc/types").RpcHandler} RpcHandler
@@ -266,35 +267,6 @@ async function keepHighlightingPlayingAction(isPopupLoad)
   {
     selectFirstAction();
   }
-}
-/**
- * @returns {Promise<import("../background/CBA").State | null>}
- */
-function getCbaState() {
-  return new Promise((resolve) => {
-    const uuid = uuidv4();
-    const onTimeout = () => {
-      removeRpcListener(handler);
-      resolve(null);
-    }
-    /** @type {RpcHandler} */
-    const handler = (state) => {
-      if (state.msgType === "GetStateResponse" && state.id === uuid) {
-        removeRpcListener(handler);
-        resolve(state.state);
-      }
-      clearTimeout(onTimeout);
-    };
-    setTimeout(onTimeout, 1000);
-    addRpcListener(handler);
-    sendRpcMessage({msgType: "GetState", id: uuid});
-  });
-}
-
-function uuidv4() {
-  return "10000000-1000-4000-8000-100000000000".replace(/[018]/g, c =>
-    (c ^ crypto.getRandomValues(new Uint8Array(1))[0] & 15 >> c / 4).toString(16)
-  );
 }
 
 function saveProjectsState()
